:bug: Adjust the first and last rows position when adjust the column width in attribute view
