+++ app/src/protyle/scroll/event.ts
@@ -4,7 +4,7 @@ import {fetchPost} from "../../util/fetch";
 import {onGet} from "../util/onGet";
 import {isMobile} from "../../util/functions";
 import {hasClosestBlock, hasClosestByClassName} from "../util/hasClosest";
-import {stickyScrollY} from "./stickyScroll";
+import {avScroll} from "../render/av/scroll";
 
 let getIndexTimeout: number;
 export const scrollEvent = (protyle: IProtyle, element: HTMLElement) => {
@@ -22,19 +22,7 @@ export const scrollEvent = (protyle: IProtyle, element: HTMLElement) => {
         }
 
         protyle.wysiwyg.element.querySelectorAll(".av").forEach((item: HTMLElement) => {
-            const bodyElement = item.querySelector(".av__body") as HTMLElement;
-
-            if (bodyElement) {
-                const headerElement = bodyElement.querySelector(".av__row--header") as HTMLElement;
-                const footerElement = bodyElement.querySelector(".av__row--footer") as HTMLElement;
-
-                stickyScrollY(
-                    element,
-                    bodyElement,
-                    headerElement ? [{element: headerElement}] : [],
-                    footerElement ? [{element: footerElement}] : [],
-                );
-            }
+            avScroll(element, item);
         });
 
         if (!protyle.element.classList.contains("block__edit") && !isMobile()) {
